fix(workflow): refacto draft status chip (#16762)
